Voxel size factor is now a float

+++ include/Core/Castor3D/Render/Technique/Voxelize/VoxelSceneData.hpp
@@ -63,7 +63,7 @@ namespace castor3d
 		//!\~english	The size of a ray step in voxel cone tracint.
 		//!\~french		La taille d'un pas de rayon dans le voxel cone tracing.
 		float rayStepSize{ 0.75f };
-		uint32_t voxelSizeFactor{ 1u };
+		float voxelSizeFactor{ 1.0f };
 	};
 }
